recruitai-backend-clean/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0

# AI and ML
openai==1.3.7
numpy==1.24.3
scikit-learn==1.3.2
google-generativeai==0.3.2

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Email
emails==0.6

# Image processing
Pillow==10.1.0

# Date and time
python-dateutil==2.8.2

# Utilities
python-slugify==8.0.1
validators==0.22.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production server
gunicorn==21.2.0

# Environment variables
python-dotenv==1.0.0

# Additional dependencies for production
psycopg2-binary==2.9.9

